+++ src/rdkafka_sasl_cyrus.c
@@ -57,6 +57,8 @@ static mtx_t rd_kafka_sasl_cyrus_kinit_lock;
  */
 typedef struct rd_kafka_sasl_cyrus_handle_s {
         rd_kafka_timer_t kinit_refresh_tmr;
+        rd_atomic32_t ready; /**< First kinit command has finished, or there
+                              *   is no kinit command. */
 } rd_kafka_sasl_cyrus_handle_t;
 
 /**
@@ -197,10 +199,12 @@ static ssize_t render_callback (const char *key, char *buf,
  * @locality rdkafka main thread
  */
 static int rd_kafka_sasl_cyrus_kinit_refresh (rd_kafka_t *rk) {
+        rd_kafka_sasl_cyrus_handle_t *handle = rk->rk_sasl.handle;
         int r;
         char *cmd;
         char errstr[128];
         rd_ts_t ts_start;
+        int duration;
 
         /* Build kinit refresh command line using string rendering and config */
         cmd = rd_string_render(rk->rk_conf.sasl.kinit_cmd,
@@ -225,6 +229,21 @@ static int rd_kafka_sasl_cyrus_kinit_refresh (rd_kafka_t *rk) {
         mtx_lock(&rd_kafka_sasl_cyrus_kinit_lock);
         r = system(cmd);
         mtx_unlock(&rd_kafka_sasl_cyrus_kinit_lock);
+
+        duration = (int)((rd_clock() - ts_start) / 1000);
+        if (duration > 5000)
+                rd_kafka_log(rk, LOG_WARNING, "SASLREFRESH",
+                             "Slow Kerberos ticket refresh: %dms: %s",
+                             duration, cmd);
+
+        /* Regardless of outcome from the kinit command (it can fail
+         * even if the ticket is available), we now allow broker connections. */
+        if (rd_atomic32_add(&handle->ready, 1) == 1) {
+                rd_kafka_dbg(rk, SECURITY, "SASLREFRESH",
+                             "First kinit command finished: waking up "
+                             "broker threads");
+                rd_kafka_all_brokers_wakeup(rk, RD_KAFKA_BROKER_STATE_INIT);
+        }
 
         if (r == -1) {
                 if (errno == ECHILD) {
@@ -259,8 +278,7 @@ static int rd_kafka_sasl_cyrus_kinit_refresh (rd_kafka_t *rk) {
         rd_free(cmd);
 
         rd_kafka_dbg(rk, SECURITY, "SASLREFRESH",
-                     "Kerberos ticket refreshed in %"PRId64"ms",
-                     (rd_clock() - ts_start) / 1000);
+                     "Kerberos ticket refreshed in %dms", duration);
         return 0;
 }
 
@@ -548,6 +566,19 @@ static int rd_kafka_sasl_cyrus_client_new (rd_kafka_transport_t *rktrans,
 
 
 /**
+ * @brief SASL/GSSAPI is ready when at least one kinit command has been
+ *        executed (regardless of exit status).
+ */
+static rd_bool_t rd_kafka_sasl_cyrus_ready (rd_kafka_t *rk) {
+        rd_kafka_sasl_cyrus_handle_t *handle = rk->rk_sasl.handle;
+
+        if (!handle)
+                return rd_false;
+
+        return rd_atomic32_get(&handle->ready) > 0;
+}
+
+/**
  * @brief Per-client-instance initializer
  */
 static int rd_kafka_sasl_cyrus_init (rd_kafka_t *rk,
@@ -566,8 +597,10 @@ static int rd_kafka_sasl_cyrus_init (rd_kafka_t *rk,
                              rk->rk_conf.sasl.relogin_min_time * 1000ll,
                              rd_kafka_sasl_cyrus_kinit_refresh_tmr_cb, rk);
 
-        /* Acquire or refresh ticket */
-        rd_kafka_sasl_cyrus_kinit_refresh(rk);
+        /* Kick off the timer immediately to refresh the ticket.
+         * (Timer is triggered from the main loop). */
+        rd_kafka_timer_override_once(&rk->rk_timers, &handle->kinit_refresh_tmr,
+                                     0/*immediately*/);
 
         return 0;
 }
@@ -653,5 +686,6 @@ const struct rd_kafka_sasl_provider rd_kafka_sasl_cyrus_provider = {
         .client_new    = rd_kafka_sasl_cyrus_client_new,
         .recv          = rd_kafka_sasl_cyrus_recv,
         .close         = rd_kafka_sasl_cyrus_close,
+        .ready         = rd_kafka_sasl_cyrus_ready,
         .conf_validate = rd_kafka_sasl_cyrus_conf_validate
 };